refactor: move state-sync related methods to a separate file
